+++ src/wizard/OSSelectionStep.qml
@@ -352,6 +352,9 @@ WizardStepBase {
             if (bytesTotal > 0) {
                 artifactDownloadProgressDialog.indeterminate = false
                 artifactDownloadProgressDialog.progress = bytesReceived / bytesTotal
+
+                // Update download speed calculation
+                artifactDownloadProgressDialog.updateDownloadSpeed(bytesReceived)
             } else {
                 artifactDownloadProgressDialog.indeterminate = true
             }